+++ CMakeLists.txt
@@ -159,8 +159,10 @@ set(
   src/shm_manager.h
   src/pb_exception.h
   src/pb_preferred_memory.h
-  src/pb_custom_metrics.h
-  src/pb_custom_metrics.cc
+  src/metric.h
+  src/metric.cc
+  src/metric_family.h
+  src/metric_family.cc
 )
 
 set(
@@ -179,8 +181,6 @@ set(
   src/stub_launcher.cc
   src/infer_payload.h
   src/infer_payload.cc
-  src/custom_metrics_registry.h
-  src/custom_metrics_registry.cc
 )
 
 list(APPEND
